third_party/WebKit/LayoutTests/http/tests/security/mixedContent/insecure-image-in-main-frame-expected.txt

CONSOLE WARNING: line 7: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-image.html' was loaded over HTTPS, but displayed insecure content from 'http://127.0.0.1:8080/security/resources/compass.jpg': this content should also be loaded over HTTPS.

This test opens a window that loads an insecure image. We should trigger a mixed content callback because the main frame in the window is HTTPS but is displaying insecure content.
